Move DeckSwiperView into study feature
